Initial commit based on the Colectiva Legal alert system CF template
[Problem]
I want to host WordPress on AWS so I don't have to pay WordPress's ridiculous fees. The goal is to run the MySQL
database and WordPress servers on Docker so I can deploy as many WordPress sites as I want.

Because of this, I'm going to use the Colectiva Legal alert system CF template as a starting point.

[Solution]
Copy the CF project to this new repository.