Create a WordPress CloudFormation template
[Problem]
I want to have a set of WordPress servers that are independent for my different projects e.g. writing, photography, and
providing one for my sibling.

[Solution]
Create a CloudFormation template that dynamically generates the WordPress servers as I need them. Comes with HTTPS
termination (woot!) and they all share the same underlying instance for this low traffic scenario.

[Test]
Went through instructions, deployed, and tested that each of the subdomains work.