Issue-1: Add a custom block device mapping to control root volume size
[Problem]
Originally, I though creating this device mapping would remove another EBS volume of 22GiB, which is available to the
instance. While this volume is not mounted, it is used by docker to store the layers. More information here:

https://github.com/aws/amazon-ecs-agent/issues/312

However, the size of the volume used is not specified, which makes reasoning about behavior of the instance over long
term not clear.

[Solution]
Create a custom Block Device Mapping of 8GB in the CF template.

It is also worth noting that the WordPress artifacts are all stored in the DB, which is hosted on EFS. Thus all the data
that is consuming the space on the root partition should be from system processes and logs. In the future, I should
create alarms around disk space and CPU utilization.

[Test]
Deployed, ssh'd into the instance and verified that the root volume is 8GiB.