Issue-5: Fix 503 after installation complete
[Problem]
After WordPress is installed, the root path returns a `503 Service Temporarily Unavailable`.

[Solution]
After installation, the root path is permanently moved (`301`). To where? I have no idea, because the site still goes to
the root path. But this was causing the Target Group's health check to fail, which manifested itself in the 503.

[Test]
After updating the health check to accept 301, the website succesfully loads.