Issue-6: WordPress servers do not recover after killing the instance
[Problem]
The databases for the WordPress servers should already exist in the EFS volume and be used by the Docker containers when
starting on the new EC2 instance. However, after killing the instance, the WordPress websites start in the un-configured
state i.e. /wp-admin/install.php.

[Solution]
The mount targets for the EFS volume were not being connected, thus an endpoint for the EC2 instance to mount the volume
did not exist.

[Test]
1. Updated the template to run the code to create the mount targets (it already existed)
1. Terminated the instance
1. `ssh` into the new instance and confirm the EFS volume is mounted
1. set up one of the docker WordPress instantiations
1. kill the instance
1. verify that instantation uses the configuration from the previous configuration